backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.5
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
numpy>=1.26
pandas>=2.0
scikit-learn>=1.3
statsmodels>=0.14
yfinance>=0.2
torch>=2.0
